requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
django-autocomplete-light==3.2.9
django-crispy-forms==1.6.1
django-filter==1.0.4
django-tables2==1.9.0
djangorestframework==3.6.3
idna==2.5
lxml==3.8.0
pytz==2017.2
requests==2.18.1
six==1.10.0
urllib3==1.21.1
SPARQLWrapper==1.8.0
